Project - Supervised Learning
Context
Buying and selling used phones and tablets used to be something that happened on a handful of 
online marketplace sites. But the used and refurbished device market has grown considerably over 
the past decade, and a new IDC (International Data Corporation) forecast predicts that the used 
phone market would be worth $52.7bn by 2023 with a compound annual growth rate (CAGR) of 
13.6% from 2018 to 2023. This growth can be attributed to an uptick in demand for used phones 
and tablets that offer considerable savings compared with new models.
Refurbished and used devices continue to provide cost-effective alternatives to both consumers 
and businesses that are looking to save money when purchasing one. There are plenty of other 
benefits associated with the used device market. Used and refurbished devices can be sold with 
warranties and can also be insured with proof of purchase. Third-party vendors/platforms, such as 
Verizon, Amazon, etc., provide attractive offers to customers for refurbished devices. Maximizing 
the longevity of devices through second-hand trade also reduces their environmental impact and 
helps in recycling and reducing waste. The impact of the COVID-19 outbreak may further boost 
this segment as consumers cut back on discretionary spending and buy phones and tablets only for 
immediate needs.
Objective
The rising potential of this comparatively under-the-radar market fuels the need for an ML-based 
solution to develop a dynamic pricing strategy for used and refurbished devices. ReCell, a startup 
aiming to tap the potential in this market, has hired you as a data scientist. They want you to 
analyze the data provided and build a linear regression model to predict the price of a used 
phone/tablet and identify factors that significantly influence it.